tools/requirements.txt
altgraph==0.17
autopep8==1.5.7
macholib==1.15
mysql-connector-python==8.0.26
paho-mqtt==1.5.1
prompt-toolkit==1.0.14
protobuf==3.17.3
psutil==5.8.0
pycodestyle==2.7.0
Pygments==2.9.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pynput==1.6.8
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Quartz==7.3
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
toml==0.10.2
wcwidth==0.2.5
